refactor(shared): move the pairing audit record beside the store both processes drive (#283)
The audit shape lived in `packages/core` when the daemon was the only thing that wrote it. `actana pair revoke`
(#283) is the second writer and it runs in the `actana` CLI, which never loads `@actana/core` — one binary does both
jobs by keeping the daemon's bundle out of its own. So the module moves to `packages/shared`, next to the pairing
store, which is there for exactly the same reason after #288.

One shape, one field list and one redaction across both writers is the point. Two modules that agreed by convention
would drift the first time a field was added to one of them, and the field list is precisely what keeps a secret out
of a log line.

`revoked` joins the outcomes. It is on this list rather than in a log of its own because it answers the same
question the others do — who holds a credential this Core signed, and how did they get it — and an operator reading
back through a pairing's life should not have to join two logs to see it end.

`LOCAL_OPERATOR_PEER` is what a revocation's `peer` carries, since it arrives at no socket at all. A constant rather
than a blank: `peer` is the field a reader scans for where something came from, and an empty one reads as a record
whose origin was lost rather than one that never had a remote origin.

Refs #283
